Minimizando clase principal
